lab7

+++ app/Http/Controllers/ArticleController.php
@@ -1,11 +1,11 @@
 <?php
 
 namespace App\Http\Controllers;
-
-use App\Models\Article;
 use App\Models\User;
+use App\Models\Article;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Log;
 class ArticleController extends Controller
 {
     /**
@@ -27,6 +27,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', [self::class]);
         $request->validate([
             'date' => 'date',
             'name' => 'required|min:5|max:100',
@@ -69,6 +70,8 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('update', $article->id);
+        // $response = Gate::inspect('update', $article->id);
         $request->validate([
             'date' => 'date',
             'name' => 'required|min:5|max:100',
@@ -89,6 +92,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', [self::class]);
         if ($article->delete())
             return redirect('/article')->with('status', value: 'Delete succsess!');
         else
